change dir structure as utils => lib
