+++ src/components/general/Date.tsx
@@ -1,0 +1,6 @@
+import { parseISO, format } from 'date-fns'
+
+export function Date({ dateString }: { dateString: string }) {
+  const date = parseISO(dateString)
+  return <time dateTime={dateString}>{format(date, 'LLL d, yyyy')}</time>
+}
