+++ src/lib/mdParser.js
